+++ CHANGELOG.md
@@ -8,6 +8,8 @@ most recent changes however.
 
 _Unreleased_
 
+- Support auto fix requires-python when conflict. #160
+
 - Added support for custom indexes.  #199
 
 - `rye add` no longer complains when a local version information is